clearing warning in the project stackoverflow https://stackoverflow.com/questions/50399090/use-of-a-signed-integer-operand-with-a-binary-bitwise-operator-when-using-un

+++ displayImageMatrix/main.cpp
@@ -4,11 +4,11 @@
 
 using namespace cv;
 
-std::string type2str(int type) {
+std::string type2str(uint type) {
     std::string r;
 
-    uchar depth = type & CV_MAT_DEPTH_MASK;
-    uchar chans = 1 + (type >> CV_CN_SHIFT);
+    uchar depth = type & (unsigned int) CV_MAT_DEPTH_MASK;
+    uchar chans = 1 + (type >> (unsigned int) CV_CN_SHIFT);
 
     switch (depth) {
         case CV_8U:
